// TCC-App/src/app/models/Evento.ts
import { Lote } from "./Lote";
import { RedeSocial } from "./RedeSocial";
import { Palestrante } from "./Palestrante";

export class Evento {    
    constructor() {}
    id: number; 
    local: string; 
    dataEvento: Date; 
    tema: string;
    qtdPessoas: number; 
    imagemURL: string; 
    telefone: string;
    email: string;
    lotes: Lote[];
    redesSociais: RedeSocial[];
    palestranteEventos: Palestrante[];
}